+++ core/iwasm/common/wasm_memory.c
@@ -147,7 +147,6 @@ wasm_runtime_free_internal(void *ptr)
 void *
 wasm_runtime_malloc(unsigned int size)
 {
-#if 0
     if (size == 0) {
         LOG_WARNING("warning: wasm_runtime_malloc with size zero\n");
         /* At lease alloc 1 byte to avoid malloc failed */
@@ -155,17 +154,12 @@ wasm_runtime_malloc(unsigned int size)
     }
 
     return wasm_runtime_malloc_internal(size);
-#endif
-    return wasm_gc_malloc(size);
 }
 
 void *
 wasm_runtime_realloc(void *ptr, unsigned int size)
 {
-#if 0
     return wasm_runtime_realloc_internal(ptr, size);
-#endif
-    return NULL;
 }
 
 void
@@ -174,18 +168,18 @@ wasm_runtime_free(void *ptr)
 #if 0
   wasm_runtime_free_internal(ptr);
 #endif
-  wasm_gc_free(ptr);
+    wasm_gc_free(ptr);
 }
 
 #if WASM_ENABLE_GC != 0
 static inline void *
-wasm_gc_malloc_internal(unsigned int size)
+wasm_gc_malloc_internal(void *heap_handle, unsigned int size)
 {
-    return mem_allocator_malloc_with_gc(pool_allocator, size);
+    return mem_allocator_malloc_with_gc(heap_handle, size);
 }
- 
+
 void *
-wasm_gc_malloc(unsigned int size)
+wasm_gc_malloc(void *heap_handle, unsigned int size)
 {
     if (size == 0) {
         LOG_WARNING("warning: wasm_runtime_malloc with size zero\n");
@@ -193,7 +187,7 @@ wasm_gc_malloc(unsigned int size)
         size = 1;
     }
 
-    return wasm_gc_malloc_internal(size);
+    return wasm_gc_malloc_internal(heap_handle, size);
 }
 
 #if WASM_GC_MANUALLY != 0
